+++ src/main/java/kz/jarkyn/backend/counterparty/model/dto/AccountResponse.java
@@ -1,12 +1,13 @@
 package kz.jarkyn.backend.counterparty.model.dto;
 
 import kz.jarkyn.backend.core.model.dto.IdNamedDto;
-import kz.jarkyn.backend.counterparty.model.CounterpartyEntity;
 import org.immutables.value.Value;
+
+import java.math.BigDecimal;
 
 @Value.Immutable
 public interface AccountResponse extends IdNamedDto {
-    Integer getBalance();
+    BigDecimal getBalance();
     String getBank();
     String getGiro();
 }
