+++ src/main/java/kz/jarkyn/backend/counterparty/model/dto/CustomerListResponse.java
@@ -3,6 +3,7 @@ package kz.jarkyn.backend.counterparty.model.dto;
 import kz.jarkyn.backend.core.model.dto.IdNamedDto;
 import org.immutables.value.Value;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Value.Immutable
@@ -10,7 +11,7 @@ public interface CustomerListResponse extends IdNamedDto {
     String getPhoneNumber();
     String getShippingAddress();
     Integer getDiscount();
-    Integer getBalance();
+    BigDecimal getBalance();
     LocalDateTime getFirstSaleMoment();
     LocalDateTime getLastSaleMoment();
     Integer getTotalSaleCount();
