+++ src/main/java/kz/jarkyn/backend/document/core/model/dto/ItemRequest.java
@@ -4,9 +4,9 @@ package kz.jarkyn.backend.document.core.model.dto;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import jakarta.annotation.Nullable;
 import kz.jarkyn.backend.core.model.dto.IdDto;
-import kz.jarkyn.backend.document.sale.model.dto.ImmutableSaleRequest;
 import org.immutables.value.Value;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Value.Immutable
@@ -15,5 +15,5 @@ public interface ItemRequest extends IdDto {
     @Nullable UUID getId();
     IdDto getGood();
     Integer getPrice();
-    Integer getQuantity();
+    BigDecimal getQuantity();
 }
