+++ src/main/java/kz/jarkyn/backend/document/core/model/dto/ItemResponse.java
@@ -4,11 +4,13 @@ package kz.jarkyn.backend.document.core.model.dto;
 import kz.jarkyn.backend.core.model.dto.IdDto;
 import org.immutables.value.Value;
 
+import java.math.BigDecimal;
+
 @Value.Immutable
 public interface ItemResponse extends IdDto {
     GoodItemResponse getGood();
-    Integer getPrice();
+    BigDecimal getPrice();
     Integer getQuantity();
     Integer getRemain();
-    Integer getCostPrice();
+    BigDecimal getCostPrice();
 }
