+++ src/main/java/kz/jarkyn/backend/document/payment/model/dto/PaymentPurposeResponse.java
@@ -4,9 +4,11 @@ package kz.jarkyn.backend.document.payment.model.dto;
 import kz.jarkyn.backend.document.core.model.dto.DocumentResponse;
 import org.immutables.value.Value;
 
+import java.math.BigDecimal;
+
 @Value.Immutable
 public interface PaymentPurposeResponse {
     DocumentResponse getPaymentIn();
     DocumentResponse getDocument();
-    Integer getAmount();
+    BigDecimal getAmount();
 }
